Cleans up the insance building
