+++ src/main/kotlin/top/data/Parameters.kt
@@ -1,0 +1,4 @@
+package top.data
+
+object Parameters {
+}
